fixup regrtest.py to handle .liquid dirs
