--- benchmarks/text-0.11.2.3/Data/Text/Encoding.hs.cgi
@@ -1,4 +1,0 @@
-*********** Constraint Information ***********
-*********** KVar Distribution *****************
-[LetE : 45, LamE : 109, PredInstE : 21, TypeInstE : 290, RecBindE : 23]
-Recursive binders: 4
