--- tests/neg/Class40.hs
@@ -1,15 +1,0 @@
-{-# LANGUAGE ScopedTypeVariables #-}
-{-@ LIQUID "--no-termination" @-}
-
-module Class where
-
-import Language.Haskell.Liquid.Prelude
-
-class Frog a where
-  mkInt :: a -> Int
-  mkInt _ = liquidAssert (0 > 1) 10
-
-
-instance Frog Int where
-
-{-@ mkInt :: (Frog a) => a -> Int @-}
